+++ semcore/feedback-form/src/FeedbackForm.jsx
@@ -71,7 +71,6 @@ class FeedbackForm extends Component {
                 ref={forwardRef}
                 {...other}
                 onSubmit={api.handleSubmit}
-                title='This is feedback'
               >
                 {typeof Children.origin === 'function' ? Children.origin(api) : <Children />}
               </SFeedbackForm>
@@ -107,7 +106,7 @@ function Cancel(props) {
 }
 
 function Notice(props) {
-  const { styles, theme = 'bg-secondary-neutral', use = 'secondary' } = props;
+  const { styles, theme = 'muted', use = 'secondary' } = props;
   const SNotice = Root;
   return sstyled(styles)(<SNotice render={NoticeSmart} theme={theme} use={use} />);
 }