+++ semcore/modal/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [4.40.0] - 2024-07-05
+
+### Changed
+
+- Version minor update due to children dependencies update (`@semcore/icon` [4.39.1 ~> 4.41.0]).
+
 ## [4.39.0] - 2024-06-26
 
 ### Changed