+++ semcore/select/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [4.44.3] - 2024-07-05
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/base-trigger` [4.39.0 ~> 4.40.0], `@semcore/dropdown-menu` [4.34.2 ~> 4.34.3], `@semcore/popper` [5.37.0 ~> 5.37.1], `@semcore/icon` [4.39.1 ~> 4.41.0]).
+
 ## [4.44.2] - 2024-06-26
 
 ### Changed