+++ semcore/time-picker/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [4.44.3] - 2024-07-05
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/button` [5.31.0 ~> 5.31.1], `@semcore/select` [4.44.2 ~> 4.44.3]).
+
 ## [4.44.2] - 2024-06-26
 
 ### Changed